convert to uint64

+++ src/hammer/hammer.cc
@@ -7,7 +7,7 @@
 void flush_row(uint8_t *row) {
     // 64 byte cache line
     for (uint32_t index = 0; index < ROW_SIZE; index += 64) {
-        clflush(row[index]);
+        clflush((uint64_t)row[index]);
     }
 }
 
@@ -41,7 +41,7 @@ uint32_t hammer_addresses(uint64_t vict_virt_addr, uint64_t attacker_virt_addr_1
     }
 
     flush_row(vict_virt_addr_ptr);
-    
+
     uint32_t number_of_bitflips_in_target = 0;
     for (uint32_t index = 0; index < ROW_SIZE; index++) {
         if (vict_virt_addr_ptr[index] != 0x55) {
